refactor: InstructorLayout에서 container 분리
